tests: relocate 10 sibling-only test files; tag heavy registry-server fuzz nightly
Files exercising ONLY sibling-module code moved to their owning sibling
repos so tests live where the code lives.

Moved to siblings (10 files, ~3600 LOC):
  webhook/      ← zz_webhook_reliability_test.go, zz_fuzz_webhook_test.go
  dataexchange/ ← zz_frame_test.go, zz_fuzz_frames_test.go
  eventstream/  ← zz_event_wire_test.go
  nameserver/   ← zz_fuzz_nameserver_test.go, zz_fuzz_nameserver_records_test.go
                  (4 funcs renamed _Web4 to coexist with sibling's own coverage)
  gateway/      ← zz_gateway_test.go, zz_fuzz_gateway_test.go
  beacon/       ← zz_fuzz_beacon_test.go

Tagged //go:build nightly (stays in web4, runs only in nightly.yml):
  tests/zz_fuzz_registry_server_test.go (1291→1263 LOC, 46 fuzz/test funcs)
    Coupled to web4 testenv (setClientSigner + TestAdminToken).

New default-tag helper files (so non-nightly consumers keep resolving):
  tests/zz_test_server_helpers_test.go:
    startTestServer + startTestServerWithStore (used by
    zz_security_phase2_test.go + zz_security_fixes_test.go)
  tests/zz_test_beacon_helpers_test.go:
    startTestBeacon + sendUDP (used by zz_security_phase2_test.go)
  Both helpers were previously inside zz_fuzz_beacon_test.go /
  zz_fuzz_registry_server_test.go — the move/tag broke the consumers
  until extracted.

Stayed in web4 (real daemon integration, need TestEnv):
  zz_webhook_test.go + zz_webhook_capture_replay_test.go — import pkg/daemon.

Verified locally before push:
  - per sibling: go test ./... -count=1 → PASS
  - web4: go vet ./tests/... → clean (no undefined symbols)
  - web4: go build ./tests/... → clean (default tags)
  - web4: go build -tags nightly ./tests/... → clean (nightly tag)
  - all 6 modified siblings already merged + green on their own main

Net web4 impact: 122 → 102 test files (~10% PR-CI scope cut on top
of Phase-1's //go:build nightly cut).
